ws63_pac

Module generic

Common register and bit access and modify traits

Structs

Periph

Generic peripheral accessor

Range

Marker for field writers are safe to write in specified inclusive range

RangeFrom

Marker for field writers are safe to write in specified inclusive range

RangeTo

Marker for field writers are safe to write in specified inclusive range

Reg

This structure provides volatile access to registers.

Safe

Marker for register/field writers which can take any value of specified width

Unsafe

You should check that value is allowed to pass to register/field writer marked with this

Traits

FieldSpec

Raw field type

IsEnum

Marker for fields with fixed values

RawReg

Raw register type (u8, u16, u32, …)

Readable

Trait implemented by readable registers to enable the read method.

RegisterSpec

Raw register type

Resettable

Reset value of the register.

Writable

Trait implemented by writeable registers.

Type Aliases

BitReader

Bit-wise field reader

BitWriter

Bit-wise write field proxy

BitWriter0C

Bit-wise write field proxy

BitWriter0S

Bit-wise write field proxy

BitWriter0T

Bit-wise write field proxy

BitWriter1C

Bit-wise write field proxy

BitWriter1S

Bit-wise write field proxy

BitWriter1T

Bit-wise write field proxy

FieldReader

Field reader.

FieldWriter

Write field Proxy

R

Register reader.

W

Register writer.