ws63_pac::generic

Trait Readable

pub trait Readable: RegisterSpec { }

Trait implemented by readable registers to enable the read method.

Registers marked with Writable can be also be modify’ed.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".