ws63_pac::generic

Trait Resettable

pub trait Resettable: RegisterSpec {
    const RESET_VALUE: Self::Ux = <Self::Ux>::ZERO;

    // Provided method
    fn reset_value() -> Self::Ux { ... }
}

Reset value of the register.

This value is the initial value for the write method. It can also be directly written to the register by using the reset method.

Provided Associated Constants

const RESET_VALUE: Self::Ux = <Self::Ux>::ZERO

Reset value of the register.

Provided Methods

fn reset_value() -> Self::Ux

Reset value of the register.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".