ws63_pac::riscv_patch

Module flpctrl

Instruction patch controller configuration

Structs

FlpctrlSpec

Instruction patch controller configuration

Type Aliases

EnableR

Field enable reader - Enable instruction patch matching

EnableW

Field enable writer - Enable instruction patch matching

LoadCompare0EnableR

Field load_compare0_enable reader - Enable load-address comparison register 0

LoadCompare0EnableW

Field load_compare0_enable writer - Enable load-address comparison register 0

LoadCompare1EnableR

Field load_compare1_enable reader - Enable load-address comparison register 1

LoadCompare1EnableW

Field load_compare1_enable writer - Enable load-address comparison register 1

Outside1mR

Field outside_1m reader - Select long-jump remap entries whose destinations may be outside the source 1 MiB region

Outside1mW

Field outside_1m writer - Select long-jump remap entries whose destinations may be outside the source 1 MiB region

R

Register FLPCTRL reader

W

Register FLPCTRL writer

WriteProtectR

Field write_protect reader - Protect the comparison registers from writes

WriteProtectW

Field write_protect writer - Protect the comparison registers from writes